SummaryReport has a list of different metrics, specific metrics still to be decided.
