+++ src/main/java/use_case/statistics/MetricType.java
@@ -1,0 +1,11 @@
+package use_case.statistics;
+
+public enum MetricType {
+    MEAN,
+    MEDIAN,
+    MODE,
+    MIN,
+    MAX,
+    COUNT,
+    STANDARD_DEVIATION
+}
